Move snapshot chunk definitions to single file
